+++ src/redturtle/prenotazioni/events/events_logger.py
@@ -3,6 +3,7 @@ from json import dumps
 from time import time
 
 from plone import api
+
 from redturtle.prenotazioni import logger